Move OptionBaseInspectionTests for QuickFixes into relevant files under RD.QuickFixes
